+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+[1.4.0]: https://github.com/johnarp/galactas-guide/releases/tag/v1.4.0
 [1.3.1]: https://github.com/johnarp/galactas-guide/releases/tag/v1.3.1
 [1.3.0]: https://github.com/johnarp/galactas-guide/releases/tag/v1.3.0
 [1.2.0]: https://github.com/johnarp/galactas-guide/releases/tag/v1.2.0
@@ -7,6 +8,19 @@
 [1.1.0]: https://github.com/johnarp/galactas-guide/releases/tag/v1.1.0
 [1.0.1]: https://github.com/johnarp/galactas-guide/releases/tag/v1.0.1
 [1.0.0]: https://github.com/johnarp/galactas-guide/releases/tag/v1.0.0
+
+## [1.4.0] - 2026-05-09
+
+### Added
+
+- Costume customization
+- One costume of each rarity with its icon for every Hero
+- Rarity card background mode
+- Option to disable hover images
+
+### Changed
+
+- Structure of [CREDITS](./CREDITS.md)
 
 ## [1.3.1] - 2026-05-05